config: add configuration for PR merge method
The GitHub pull request API support squash, rebase, and merge. As
explored in issue #195, with signed commits required on the target
branch, it's not possible to use the rebase due to a known GitHub
implementation issue.

This change adds a new per-repo string configuration attribute,
`mergeMethod`, `"rebase"` by default. The current behaviour is captured by
this default value.

commit-id:b4ef58dc